requirements.txt
accelerate==1.2.0
astropy==6.1.7
astropy-iers-data==0.2025.3.10.0.29.26
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comet-ml==3.47.4
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
diffusers==0.31.0
dulwich==0.22.6
einops==0.8.1
ema-pytorch==0.7.7
everett==3.1.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
gdown==5.2.0
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.0.0
polars==1.25.2
psutil==6.1.0
pyerfa==2.0.1.5
Pygments==2.18.0
pyparsing==3.2.0
python-box==6.1.0
python-dateutil==2.9.0.post0
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rocket-fft==0.2.5
rpds-py==0.22.3
safetensors==0.4.5
scipy==1.15.2
semantic-version==2.10.0
sentry-sdk==2.19.2
simplejson==3.19.3
six==1.17.0
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.47.0
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
wurlitzer==3.1.1
xformers==0.0.29.post3
zipp==3.21.0